+++ bigquery_datatransfer/synth.py
@@ -62,7 +62,7 @@ s.replace(
 # ----------------------------------------------------------------------------
 # Add templated files
 # ----------------------------------------------------------------------------
-templated_files = common.py_library(unit_cov_level=79, cov_level=79)
+templated_files = common.py_library(unit_cov_level=79, cov_level=79, samples_test=True)
 s.move(templated_files)
 
 s.shell.run(["nox", "-s", "blacken"], hide_output=False)